UHF-7486: UHF-7486: Separate daycare form to separate file
